pyproject.toml
[tool.poetry]
name = "alex_leontiev_toolbox_python"
version = "0.1.0"
description = ""
authors = ["Oleksii Leontiev <alozz1991@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
google-cloud-bigquery = "^3.2.0"
sqlparse = "^0.4.2"
SQLAlchemy = "^1.4.39"
pytest = "^7.1.2"
Jinja2 = "^3.1.2"
pandas = "^1.4.3"
db-dtypes = "^1.0.2"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
